docs: fix gvalue.Or and adjust gsync example test
