+++ stdwrap/gsync/pool_example_test.go
@@ -18,30 +18,6 @@ import (
 	"fmt"
 )
 
-func ExampleMap() {
-	sm := Map[string, int]{}
-	sm.Store("k", 1)
-	fmt.Println(sm.Load("k"))          // 1 true
-	fmt.Println(sm.LoadO("k").Value()) // 1
-	sm.Store("k", 2)
-	fmt.Println(sm.Load("k"))           // 2 true
-	fmt.Println(sm.LoadAndDelete("k"))  // 2 true
-	fmt.Println(sm.Load("k"))           // 0 false
-	fmt.Println(sm.LoadOrStore("k", 3)) // 3 false
-	fmt.Println(sm.Load("k"))           // 3 true
-	fmt.Println(sm.ToMap())             // {"k":3}
-
-	// Output:
-	// 1 true
-	// 1
-	// 2 true
-	// 2 true
-	// 0 false
-	// 3 false
-	// 3 true
-	// map[k:3]
-}
-
 func ExamplePool() {
 	pool := Pool[*int]{
 		New: func() *int {
